AUT-5319: Move Access Token generation & signing classes to shared
- Access token generation needs to happen in both the auth internal
  API and auth external API to facilitate providing Home an AMC
  Access Token via Orchestration. Because of this, Access Token
  generation classes have been moved to shared
- All of the moved classes will be multi-domain anyway after
  changes following this commit, e.g. used in both the auth
  internal API and auth external API
- The AccountManagementScope class has not been moved as it is
  to be used only in the auth internal API
