+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/entity/amc/AMCJourneyType.java
@@ -1,5 +1,6 @@
 package uk.gov.di.authentication.frontendapi.entity.amc;
 
+import uk.gov.di.authentication.shared.entity.AccountDataScope;
 import uk.gov.di.authentication.shared.services.ConfigurationService;
 
 import java.util.List;
