+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/entity/amc/AccessTokenConfig.java
@@ -1,4 +1,6 @@
 package uk.gov.di.authentication.frontendapi.entity.amc;
 
+import uk.gov.di.authentication.shared.entity.AccessTokenScope;
+
 public record AccessTokenConfig(
         String accessTokenName, AccessTokenScope scope, String audience, String signingKey) {}
