+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/entity/amc/AccountManagementScope.java
@@ -1,5 +1,7 @@
 package uk.gov.di.authentication.frontendapi.entity.amc;
 
+import uk.gov.di.authentication.shared.entity.AccessTokenScope;
+
 public enum AccountManagementScope implements AccessTokenScope {
     ACCOUNT_DELETE("account-delete");
