+++ shared/src/test/java/uk/gov/di/authentication/shared/services/AccessTokenConstructorServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.frontendapi.services;
+package uk.gov.di.authentication.shared.services;
 
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JOSEObjectType;
@@ -16,12 +16,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import uk.gov.di.authentication.frontendapi.entity.JwtFailureReason;
-import uk.gov.di.authentication.frontendapi.entity.amc.AccessTokenScope;
-import uk.gov.di.authentication.frontendapi.entity.amc.AccountDataScope;
-import uk.gov.di.authentication.frontendapi.entity.amc.AccountManagementScope;
+import uk.gov.di.authentication.shared.entity.AccessTokenScope;
+import uk.gov.di.authentication.shared.entity.AccountDataScope;
+import uk.gov.di.authentication.shared.entity.JwtFailureReason;
 import uk.gov.di.authentication.shared.entity.Result;
-import uk.gov.di.authentication.shared.services.ConfigurationService;
 
 import java.text.ParseException;
 import java.time.Instant;
@@ -68,7 +66,7 @@ class AccessTokenConstructorServiceTest {
     private static Stream<Arguments> validScopes() {
         return Stream.of(
                 Arguments.of(AccountDataScope.PASSKEY_CREATE, "passkey-create"),
-                Arguments.of(AccountManagementScope.ACCOUNT_DELETE, "account-delete"));
+                Arguments.of(AccountDataScope.PASSKEY_RETRIEVE, "passkey-retrieve"));
     }
 
     @ParameterizedTest
